XMLUserDatabase does not allow empty wiki names
jdom returns an empty string (that is, not null) for missing attributes, so checking for null at getWikiNames results in unreachable code. Not sure if it's intended or a bug, but given that the UI on the registration form mandates a not empty value for the wiki name, let's follow that when returning wiki names.

Also, applied code formatting