JSPWIKI-388 - ParamTag loops forever when parent tag is not ParamHandler
This commit addresses an issue in the ParamTag class where, if the parent tag was not an instance of ParamHandler, an infinite loop would occur. This was due to the 'getParent()' method always returning the same object.

The fix involves getting the parent of the current tag in the loop, thereby ensuring that we keep moving up the tree until we find a parent that is a ParamHandler or until we run out of parents. This prevents the possibility of an infinite loop and provides a more robust handling of the tag structure.